CW RMM - Windows 11 compatibility solution
CW RMM - Windows 11 compatibility solution

This includes the Windows 11 validation on Windows 11 machine as well due to:
I couldn't find anything concrete that 100% validates that devices will not go to 24H2 if they are not compatible. However, everything always pointed back to checking the system requirements. So I think that might confirmation enough for us to say that Microsoft will not allow machines to 24H2 unless they are compatible.

I do believe that this could be skirted if you installed the ISO locally. But that is not something we can do, obviously.

Hence, auditing on Windows 11 will also help to find issued Windows 11 machines that gets corrupted or had issues in further upgrades if forcefully updated to 11.