Automate users permission audit
This script creates a custom table `user-class-mapping` at the very first stage and then runs at every 2 hour interval and compare the Automate user's permission change by storing it to the temporary table and match with the stored custom table data.
If the changes found then it creates the ticket with the details.